Don't minimise trauma
Saying that sexual assault "sometimes" causes trauma is pretty offensive
to those of us who *have* experienced sexual assault. The traumas may
not always be noticable, even to victims, but they're always there.

The word "sometimes" also puts an additional burden on victims because
when people think that trauma is just an occasional possibility, there
are many people who them assume that the victim doesn't have trauma,
usually for completely superficial reasons like "oh they seemed fine to
me when we chatted in that PR". It then becomes an additional, very
difficult burden for a victim to need to summon the courage and strength
to bring up and justify their trauma in certain situations.

We have much greater understanding now of PTSD and CPTSD. Let's
demonstrate that here and reduce the burden upon victims.